+++ ElvUI_MerathilisUI/Core/Changelog/7.02.lua
@@ -1,8 +1,15 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[702] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
+	RELEASE_DATE = "06.12.2025",
+	FIXES = {
+		"[Tooltip Option]: Fix that the option wasn't visble at all",
+		"[Install]: Fix an error in the Chat Install step",
+	},
 	NEW = {},
-	IMPROVEMENTS = {},
+	IMPROVEMENTS = {
+		"[Install]: Add HousingControlFrameMover to the install",
+		"[Vehicle Bar]: Remove Vigor Bar & Speed Text",
+		"[Skins]: Remove HomeBound skin",
+	},
 }