+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ### Changes ###
 
-  * [Core]: Fix a typo in the database conversation that broke my Plugin.
+  * [Tooltip Option]: Fix that the option wasn't visble at all
+  * [Install]: Fix an error in the Chat Install step
+  * [Install]: Add Housing Control Mover settings to the install
+  * [Vehicle Bar]: Remove Vigor Bar & Speed Text
+  * [Skins]: Remove HouseBound skin